Enabling UnsafeTypeCastWarningLevel for replay streamers
#jira none
#preflight 6418f7d83f3d31c94ac66da7
#rb ryan.gerleve

[CL 24732060 by Brian Bekich in ue5-main branch]

+++ Engine/Source/Runtime/NetworkReplayStreaming/InMemoryNetworkReplayStreaming/Public/InMemoryNetworkReplayStreaming.h
@@ -97,16 +97,16 @@ public:
 		, Chunks(InChunks)
 	{}
 
-	virtual void	Serialize( void* V, int64 Length );
+	virtual void	Serialize(void* V, int64 Length);
 	virtual int64	Tell();
 	virtual int64	TotalSize();
-	virtual void	Seek( int64 InPos );
+	virtual void	Seek(int64 InPos);
 	virtual bool	AtEnd();
 
 private:
 	FInMemoryReplay::FStreamChunk* GetCurrentChunk() const;
 
-	int32 Pos;
+	int64 Pos;
 	TArray<FInMemoryReplay::FStreamChunk>& Chunks;
 };
